requirements.txt
# Slice 1 Requirements - Minimal dependencies
streamlit>=1.28.0
requests>=2.31.0
rank-bm25>=0.2.2
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.5.0

# Future slices (commented out for now)
# Slice 3 - Semantic search
# sentence-transformers>=2.2.0
# faiss-cpu>=1.7.4

# Slice 4 - Knowledge graph
# spacy>=3.7.0
# networkx>=3.1
# neo4j>=5.14.0